MCU8MASS-521, MCU8MASS-522 Update inject script for cryptoauth lib
This commit changes the inject script to clone cryptoauthlib from github (.gitignore is changed to ignore the repo once cloned). It also fixes proper support for the 48 pin variant with passing mmcu flag from the CMakeList file. F_CPU is also set for compiling cryptoauthlib. This fixes the bug in the hal_driver where a delay had to be done before retrieving messages.